ParamSplitter: created a new parser class for argument splitting
It will return the next subparameter delimited by a give separator (i.e.
character) with call to next_param(). We can also check if we have
reached the end of subparameters by calling reached_end().
If next_param() is called after reaching the end of the subparameters,
an empty string will be returned.